renamed pkgs and added license headers
