--- src/main/java/com/example/gsoc/controller/PluginController.java
@@ -1,11 +1,0 @@
-package com.example.gsoc.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class PluginController {
-	
-	public PluginController() {
-	}
-
-}
