--- src/main/java/com/example/gsoc/repository/PluginRepository.java
@@ -1,7 +1,0 @@
-package com.example.gsoc.repository;
-import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.gsoc.model.Plugin;
-
-public interface PluginRepository extends JpaRepository<Plugin, Long> {
-
-}
