--- src/test/java/com/example/gsoc/GsocApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.gsoc;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GsocApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
